examples/requirements.txt
numpy
openai
pydantic
faiss-cpu
